third_party/blink/web_tests/platform/generic/fast/dom/webtiming-document-open-expected.txt

This test verifies that the NavigationTimings don't change after a document.open().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS timing.connectEnd is originalTiming.connectEnd
PASS timing.connectStart is originalTiming.connectStart
PASS timing.domComplete is originalTiming.domComplete
PASS timing.domContentLoadedEventEnd is originalTiming.domContentLoadedEventEnd
PASS timing.domContentLoadedEventStart is originalTiming.domContentLoadedEventStart
PASS timing.domInteractive is originalTiming.domInteractive
PASS timing.domLoading is originalTiming.domLoading
PASS timing.domainLookupEnd is originalTiming.domainLookupEnd
PASS timing.domainLookupStart is originalTiming.domainLookupStart
PASS timing.fetchStart is originalTiming.fetchStart
PASS timing.loadEventEnd is originalTiming.loadEventEnd
PASS timing.loadEventStart is originalTiming.loadEventStart
PASS timing.navigationStart is originalTiming.navigationStart
PASS timing.redirectEnd is originalTiming.redirectEnd
PASS timing.redirectStart is originalTiming.redirectStart
PASS timing.requestStart is originalTiming.requestStart
PASS timing.responseEnd is originalTiming.responseEnd
PASS timing.responseStart is originalTiming.responseStart
PASS timing.secureConnectionStart is originalTiming.secureConnectionStart
PASS timing.toJSON is originalTiming.toJSON
PASS timing.unloadEventEnd is originalTiming.unloadEventEnd
PASS timing.unloadEventStart is originalTiming.unloadEventStart
PASS successfullyParsed is true

TEST COMPLETE
